[chore] Tidy code for modern Java versions
Addresses some deprecations in JRuby that have been removed in 10.1.

(cherry picked from commit d6336323 - excludes generics change to interfaces)

+++ src/main/java/org/jruby/rack/DefaultRackDispatcher.java
@@ -55,7 +55,7 @@ public class DefaultRackDispatcher extends AbstractRackDispatcher {
             // TODO seems redundant maybe we should let the container decide ?!
             context.log(RackLogger.ERROR, "error app failed to handle exception: " + e, re);
             Integer errorCode = getErrorApplicationFailureStatusCode();
-            if ( errorCode != null && errorCode.intValue() > 0 ) {
+            if ( errorCode != null && errorCode > 0 ) {
                 response.sendError(errorCode);
             }
             else {
